requirements.txt
cachetools==5.5.0
certifi==2024.6.2
charset-normalizer==3.3.2
colorama==0.4.6
Deprecated==1.2.14
dnspython==2.6.1
google-api-core==2.19.2
google-auth==2.34.0
google-cloud-appengine-logging==1.4.5
google-cloud-audit-log==0.3.0
google-cloud-core==2.4.1
google-cloud-logging==3.11.2
googleapis-common-protos==1.65.0
grpc-google-iam-v1==0.13.1
grpcio==1.66.1
grpcio-status==1.66.1
idna==3.7
importlib_metadata==8.4.0
# MetaTrader5==5.0.4288
# mt5linux==0.1.9
numpy==1.26.4
opentelemetry-api==1.27.0
pandas==2.2.2
pandas_ta==0.3.14b0
proto-plus==1.24.0
protobuf==5.28.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pymongo==4.7.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
rsa==4.9
six==1.16.0
tqdm==4.66.4
tzdata==2024.1
urllib3==2.2.1
wrapt==1.16.0
zipp==3.20.2
ta-lib==0.5.2
psycopg2
logtail-python
